fix(catalog/docs): corregir firma de priceWithTax y limpiar conteos de tests
CIERRE DE P3-002:
El método `Product::priceWithTax()` declaraba `: float` como tipo de retorno,
aunque internamente ya realizaba un cast a `(int)`. Esto contradecía ADR-018.
- Solución: Cambiar firma a `public function priceWithTax(): int`.

CIERRE DE P3-003:
La documentación (README.md, PROJECT_STATUS.md, ADRs) contenía números
hardcodeados de tests (1,469, 1016, 453, 432) que generan discrepancias
con el estado real del repositorio.
- Solución: Reemplazar conteos estáticos por instrucciones dinámicas
  ("Verificar vía CI" o comandos locales), eliminando la deuda de
  documentación desactualizada.

IMPACTO:
✅ Coherencia total de tipos en el dominio de Catálogo (ADR-018).
✅ Documentación libre de números hardcodeados que requieren mantenimiento manual.
✅ Prevención de confusiones sobre el estado real de la suite de tests.

Refs: P3-002 Product::priceWithTax float, P3-003 Doc test count discrepancies

+++ app/Modules/Catalog/Domain/Entities/Product.php
@@ -210,7 +210,7 @@ class Product extends Model
      * Usa la misma fuente que OrderItem::saving() para garantizar consistencia:
      * getEffectiveTax() → product.tax_id → category.tax_id → company default → tax_rate legacy.
      */
-    public function priceWithTax(): float
+    public function priceWithTax(): int
     {
         $effectiveTax = $this->getEffectiveTax();
         $rate = $effectiveTax ? (float) $effectiveTax->rate : (float) ($this->tax_rate ?? 0);
